+++ packages/sdk-plugin-tailordb-erd/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @tailor-platform/sdk-tailordb-erd-plugin
 
+## 0.2.3
+
+### Patch Changes
+
+- [#2332](https://github.com/tailor-platform/sdk/pull/2332) [`766a3f9`](https://github.com/tailor-platform/sdk/commit/766a3f91acaf7fa9cd643f7c2bf9fe6055d578c5) Thanks [@renovate](https://github.com/apps/renovate)! - fix(deps): update dependency @clack/prompts to v1.8.0
+
+- [#2378](https://github.com/tailor-platform/sdk/pull/2378) [`867d586`](https://github.com/tailor-platform/sdk/commit/867d586a8b08a27f0b1a26dec2f18efa7963e375) Thanks [@renovate](https://github.com/apps/renovate)! - fix(deps): update dependency zod to v4.6.5
+- Updated dependencies [[`f5a3ad8`](https://github.com/tailor-platform/sdk/commit/f5a3ad8212dc5eec1ce54265d7a841aea96c7944), [`766a3f9`](https://github.com/tailor-platform/sdk/commit/766a3f91acaf7fa9cd643f7c2bf9fe6055d578c5), [`867d586`](https://github.com/tailor-platform/sdk/commit/867d586a8b08a27f0b1a26dec2f18efa7963e375), [`8d65e44`](https://github.com/tailor-platform/sdk/commit/8d65e44c5f4462637a9f3b60a2c0241b8edbb098)]:
+  - @tailor-platform/sdk@2.19.0
+
 ## 0.2.2
 
 ### Patch Changes